Do not rewrite logs in rescue mode
When running in the rescue mode, anaconda should not rewrite
installation logs on the system being rescued with the log files from
the rescue environment.
Also do not change any SELinux contexts automatically on the system
being rescued.

Resolves: RHEL-4794

+++ data/post-scripts/99-copy-logs.ks
@@ -7,6 +7,13 @@ NOSAVE_INPUT_KS_FILE=/tmp/NOSAVE_INPUT_KS
 NOSAVE_LOGS_FILE=/tmp/NOSAVE_LOGS
 PRE_ANA_LOGS=/tmp/pre-anaconda-logs
 DNF_DEBUG_LOGS=/root/debugdata
+RESCUE_MODE=/tmp/RESCUE_MODE
+
+# Do not copy log files from the rescue environment to the system being rescued to avoid
+# rewriting logs from the original installation of the system.
+if [ -e ${RESCUE_MODE} ]; then
+    exit 0
+fi
 
 if [ -e ${NOSAVE_LOGS_FILE} ]; then
     rm -f ${NOSAVE_LOGS_FILE}
